requirements.txt
streamlit
requests
schedule
python-dotenv
pymongo